Add files via upload
This contains scatter graphs plotted to show how each variable relates to new_cases and new_deaths which are going to be the basis of our study. A regression line has been added where possible to better show the relationship. We are using this data in our exploratory data analysis and our feature selection along with other methods.